Move template provider code to pkg/external/customtemplates dir
